requirements.txt
deepeval==2.7.9
python-dotenv==1.0.1